Update QMK_KeymapToAsciiTable.py
This can now automatically recognize additional KeyCodes that are not hardcoded into the QMK dictionary. However, there are certain limitations to this recognitions. Layers that are used in toggles such as LT and MO must be present as numbers.

This should in theory work for all keycodes of the form [A-Z]*([0,9]) and [A-Z]*([0,9],[A-Z]*). Note however, there are keycodes such as LT(Layer, tap_key) that will unfortunately break the nicely arranged ascii layout.